src/pyproject.toml
[project]
name = "stac"
description = "stac-fastapi"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
    {name = "Vincent Sarago", email = "vincent@developmentseed.com"},
]

dynamic = ["version"]
dependencies = [
    "stac-fastapi.api @ git+https://github.com/vincentsarago/stac-fastapi/@a3c3eb1508929ebd458552373043d7d627385ec6#egg=stac-fastapi.api&subdirectory=stac_fastapi/api",
    "stac-fastapi.extensions @ git+https://github.com/vincentsarago/stac-fastapi/@a3c3eb1508929ebd458552373043d7d627385ec6#egg=stac-fastapi.extensions&subdirectory=stac_fastapi/extensions",
    "stac-fastapi.pgstac @ git+https://github.com/vincentsarago/stac-fastapi/@a3c3eb1508929ebd458552373043d7d627385ec6#egg=stac-fastapi.pgstac&subdirectory=stac_fastapi/pgstac",
    "stac-fastapi.types @ git+https://github.com/vincentsarago/stac-fastapi/@a3c3eb1508929ebd458552373043d7d627385ec6#egg=stac-fastapi.types&subdirectory=stac_fastapi/types",
    "starlette-cramjam>=0.1.0.a0,<0.2",
    "pypgstac==0.6.3",
    "psycopg",
    "psycopg_pool",
]

[build-system]
requires = ["flit>=3.2,<4"]
build-backend = "flit_core.buildapi"
